Add arcade machine with playable mini-game
Introduces an arcade machine building that can be placed in build mode. Interacting with the arcade machine opens a new 'Shiba Runner' mini-game in a dedicated UI overlay, locking player movement until exited. Includes new textures, sounds, and supporting scripts for the arcade feature, as well as UI and input updates to support entering and exiting the arcade mode.